ai-service/vision/requirements.txt

torch>=2.1.0
transformers>=4.36.0
peft>=0.7.0
pillow>=10.1.0
qwen-vl-utils>=0.0.1
bitsandbytes>=0.41.3
accelerate>=0.25.0
